pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-gateway"
version = "0.1.0"
description = "Local LLM routing gateway for Apple Silicon Mac Mini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "litellm>=1.55,<2",
    "pyyaml>=6.0,<7",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "httpx>=0.28,<1",
    "python-json-logger>=3.2,<4",
    "opentelemetry-api>=1.29,<2",
    "opentelemetry-sdk>=1.29,<2",
    "opentelemetry-exporter-otlp>=1.29,<2",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-client>=0.21,<1",
    "watchfiles>=1.0,<2",
    "tiktoken>=0.8,<1",
    "aiosqlite>=0.20,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",
    "ruff>=0.8",
]
local = [
    "mlx-lm>=0.21; sys_platform == 'darwin'",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_gateway"]

[tool.ruff]
line-length = 100
target-version = "py311"
